+++ .eslintrc.js
@@ -1,0 +1,2 @@
+// Eat own dogfood
+module.exports = require('./index');