+++ ember-addon.js
@@ -1,7 +1,5 @@
 module.exports = {
-  extends: [
-    '@zestia/eslint-config/ember'
-  ],
+  extends: ['@zestia/eslint-config/ember'],
 
   overrides: [
     // node files
@@ -30,9 +28,13 @@ module.exports = {
         node: true
       },
       plugins: ['node'],
-      rules: Object.assign({}, require('eslint-plugin-node').configs.recommended.rules, {
-        // add your custom rules and overrides for node files here
-      })
+      rules: Object.assign(
+        {},
+        require('eslint-plugin-node').configs.recommended.rules,
+        {
+          // add your custom rules and overrides for node files here
+        }
+      )
     }
   ]
 };